# backend/openapi.json
{
    "openapi": "3.0.0",
    "paths": {
        "/api/shops": {
            "get": {
                "operationId": "ShopsController_findAll",
                "parameters": [],
                "responses": {
                    "200": {
                        "description": "Successful response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/ShopDTO"
                                    }
                                }
                            }
                        }
                    },
                    "default": {
                        "description": "Erroneous response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorApiResponse"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "cookie": []
                    }
                ],
                "summary": "Get all shops",
                "tags": [
                    "shops"
                ]
            }
        },
        "/api/items/shop/{shopInternalId}": {
            "get": {
                "operationId": "ItemsController_findAllUnfulfilledByShop",
                "parameters": [
                    {
                        "name": "shopInternalId",
                        "required": true,
                        "in": "path",
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/ItemDTO"
                                    }
                                }
                            }
                        }
                    },
                    "default": {
                        "description": "Erroneous response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorApiResponse"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "cookie": []
                    }
                ],
                "summary": "Get all unfulfilled items by shop",
                "tags": [
                    "items"
                ]
            }
        },
        "/api/items": {
            "post": {
                "operationId": "ItemsController_create",
                "parameters": [],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/CreateItemDTO"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Successful response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ItemDTO"
                                }
                            }
                        }
                    },
                    "default": {
                        "description": "Erroneous response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorApiResponse"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "cookie": []
                    }
                ],
                "summary": "Create item",
                "tags": [
                    "items"
                ]
            }
        },
        "/api/items/{itemId}/buy": {
            "patch": {
                "operationId": "ItemsController_buyItem",
                "parameters": [
                    {
                        "name": "itemId",
                        "required": true,
                        "in": "path",
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/BuyItemDTO"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Successful response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ItemDTO"
                                }
                            }
                        }
                    },
                    "default": {
                        "description": "Erroneous response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorApiResponse"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "cookie": []
                    }
                ],
                "summary": "Update item bought counter",
                "tags": [
                    "items"
                ]
            }
        },
        "/api/items/{itemId}/cancel": {
            "patch": {
                "operationId": "ItemsController_cancelItem",
                "parameters": [
                    {
                        "name": "itemId",
                        "required": true,
                        "in": "path",
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ItemDTO"
                                }
                            }
                        }
                    },
                    "default": {
                        "description": "Erroneous response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorApiResponse"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "cookie": []
                    }
                ],
                "summary": "Cancel item",
                "tags": [
                    "items"
                ]
            }
        },
        "/api/api-keys": {
            "get": {
                "operationId": "ApiKeysController_findMy",
                "parameters": [],
                "responses": {
                    "200": {
                        "description": "Successful response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/ApiKeyDTO"
                                    }
                                }
                            }
                        }
                    },
                    "default": {
                        "description": "Erroneous response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorApiResponse"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "cookie": []
                    }
                ],
                "summary": "Get your API keys",
                "tags": [
                    "api-keys"
                ]
            },
            "post": {
                "operationId": "ApiKeysController_create",
                "parameters": [],
                "responses": {
                    "200": {
                        "description": "Successful response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ApiKeyDTO"
                                }
                            }
                        }
                    },
                    "default": {
                        "description": "Erroneous response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorApiResponse"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "cookie": []
                    }
                ],
                "summary": "Create new API key",
                "tags": [
                    "api-keys"
                ]
            }
        },
        "/api/api-keys/{id}": {
            "delete": {
                "operationId": "ApiKeysController_remove",
                "parameters": [
                    {
                        "name": "id",
                        "required": true,
                        "in": "path",
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/EmptyResponse"
                                }
                            }
                        }
                    },
                    "default": {
                        "description": "Erroneous response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorApiResponse"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "cookie": []
                    }
                ],
                "summary": "Delete API key",
                "tags": [
                    "api-keys"
                ]
            }
        },
        "/api/integrations/chipdip/parse": {
            "post": {
                "operationId": "ChipdipController_parse",
                "parameters": [],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/ParseChipdipDTO"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Successful response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ItemDetailsDTO"
                                }
                            }
                        }
                    },
                    "default": {
                        "description": "Erroneous response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorApiResponse"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "cookie": []
                    }
                ],
                "summary": "Parse chipdip URL/SKU",
                "tags": [
                    "integrations/chipdip"
                ]
            }
        },
        "/api/auth/self": {
            "get": {
                "operationId": "AuthController_getSelf",
                "parameters": [],
                "responses": {
                    "200": {
                        "description": "Successful response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/SelfAuthInfoDTO"
                                }
                            }
                        }
                    },
                    "default": {
                        "description": "Erroneous response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorApiResponse"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "cookie": []
                    }
                ],
                "summary": "Get self member auth info",
                "tags": [
                    "auth"
                ]
            }
        }
    },
    "info": {
        "title": "EBuyer",
        "description": "Buylist application",
        "version": "1.0",
        "contact": {}
    },
    "tags": [],
    "servers": [],
    "components": {
        "securitySchemes": {
            "cookie": {
                "type": "apiKey",
                "in": "cookie",
                "name": "session"
            }
        },
        "schemas": {
            "ShopDTO": {
                "type": "object",
                "properties": {
                    "id": {
                        "type": "string",
                        "format": "uuid"
                    },
                    "internalId": {
                        "type": "string"
                    },
                    "name": {
                        "type": "string"
                    }
                },
                "required": [
                    "id",
                    "internalId",
                    "name"
                ]
            },
            "ErrorApiResponse": {
                "type": "object",
                "properties": {
                    "statusCode": {
                        "type": "number"
                    },
                    "message": {
                        "type": "string"
                    }
                },
                "required": [
                    "statusCode",
                    "message"
                ]
            },
            "ItemDTO": {
                "type": "object",
                "properties": {
                    "id": {
                        "type": "string",
                        "format": "uuid"
                    },
                    "title": {
                        "type": "string"
                    },
                    "comment": {
                        "type": "string"
                    },
                    "sku": {
                        "type": "string"
                    },
                    "imageUrl": {
                        "type": "string"
                    },
                    "itemUrl": {
                        "type": "string"
                    },
                    "price": {
                        "type": "string"
                    },
                    "requiredCount": {
                        "type": "number"
                    },
                    "boughtCount": {
                        "type": "number"
                    },
                    "status": {
                        "type": "string",
                        "enum": [
                            "created",
                            "canceled",
                            "fulfilled"
                        ]
                    },
                    "additionalData": {
                        "type": "object"
                    }
                },
                "required": [
                    "id",
                    "title",
                    "requiredCount",
                    "boughtCount",
                    "status"
                ]
            },
            "CreateItemDTO": {
                "type": "object",
                "properties": {
                    "shopInternalId": {
                        "type": "string"
                    },
                    "title": {
                        "type": "string"
                    },
                    "comment": {
                        "type": "string"
                    },
                    "sku": {
                        "type": "string"
                    },
                    "imageUrl": {
                        "type": "string"
                    },
                    "itemUrl": {
                        "type": "string"
                    },
                    "price": {
                        "type": "string"
                    },
                    "requiredCount": {
                        "type": "number"
                    },
                    "additionalData": {
                        "type": "object"
                    }
                },
                "required": [
                    "shopInternalId",
                    "title",
                    "requiredCount"
                ]
            },
            "BuyItemDTO": {
                "type": "object",
                "properties": {
                    "additionalBoughtCount": {
                        "type": "number"
                    }
                },
                "required": [
                    "additionalBoughtCount"
                ]
            },
            "ApiKeyDTO": {
                "type": "object",
                "properties": {
                    "id": {
                        "type": "string",
                        "format": "uuid"
                    },
                    "key": {
                        "type": "string"
                    }
                },
                "required": [
                    "id",
                    "key"
                ]
            },
            "EmptyResponse": {
                "type": "object",
                "properties": {}
            },
            "ParseChipdipDTO": {
                "type": "object",
                "properties": {
                    "data": {
                        "type": "string"
                    }
                },
                "required": [
                    "data"
                ]
            },
            "ItemDetailsDTO": {
                "type": "object",
                "properties": {
                    "title": {
                        "type": "string"
                    },
                    "sku": {
                        "type": "string"
                    },
                    "imageUrl": {
                        "type": "string"
                    },
                    "itemUrl": {
                        "type": "string"
                    },
                    "price": {
                        "type": "string"
                    }
                }
            },
            "SelfAuthInfoDTO": {
                "type": "object",
                "properties": {
                    "id": {
                        "type": "string",
                        "format": "uuid"
                    }
                },
                "required": [
                    "id"
                ]
            }
        }
    }
}